feat: expose types to public api
