+++ playground/tsconfig.json
@@ -1,3 +1,8 @@
 {
-  "extends": "./.nuxt/tsconfig.json"
+  "extends": "./.nuxt/tsconfig.json",
+  "compilerOptions": {
+    "paths": {
+      "nuxt-lottie": ["../src/types.ts"]
+    }
+  }
 }
